% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmcorrShiny.R
\name{rmcorrShiny}
\alias{rmcorrShiny}
\title{A shiny application for repeated measures correlation}
\usage{
rmcorrShiny()
}
\description{
rmcorrShiny provides an intuitive interface with options to compute and plot the repeated measures correlation.
}
\examples{
if (interactive()) {
  rmcorrShiny::rmcorrShiny()
}
}
\references{
Bakdash, J.Z., & Marusich, L.R. (2017).
Repeated Measures Correlation. \emph{Frontiers in Psychology, 8}, 256.
\doi{10.3389/fpsyg.2017.00456}
}
\seealso{
Web version: \url{https://lmarusich.shinyapps.io/shiny_rmcorr/}

\link[rmcorr:rmcorr]{rmcorr::rmcorr}
}
